+++ src/mpd_client/connection.h
@@ -14,9 +14,6 @@
 #include "src/lib/mympd_state.h"
 
 bool mpd_client_connect(struct t_partition_state *partition_state);
-bool mpd_client_set_keepalive(struct t_partition_state *partition_state);
-bool mpd_client_set_timeout(struct t_partition_state *partition_state);
-bool mpd_client_set_binarylimit(struct t_partition_state *partition_state);
 bool mpd_client_set_connection_options(struct t_partition_state *partition_state);
 void mpd_client_disconnect(struct t_partition_state *partition_state);
 void mpd_client_disconnect_silent(struct t_partition_state *partition_state);